+++ hosts/hermes/darwin-configuration.nix
@@ -20,7 +20,7 @@
 
   system.defaults.dock.persistent-apps = [
     "${pkgs.firefox-bin}/Applications/Firefox.app"
-    "/Applications/Ghostty.app"
+    "${pkgs.ghostty-bin}/Applications/Ghostty.app"
     "/Applications/1Password.app"
     "${pkgs.vscode}/Applications/Visual Studio Code.app"
     "${pkgs.spotify}/Applications/Spotify.app"