+++ overlays/1password.nix
@@ -1,0 +1,8 @@
+self: super: {
+  _1password-gui = super._1password-gui.overrideAttrs (old: {
+    meta = assert old.meta.broken == super.stdenv.hostPlatform.isDarwin;
+      old.meta // {
+        broken = false;
+      };
+  });
+}