+++ overlays/mas.nix
@@ -1,0 +1,5 @@
+self: super: {
+  mas = assert !super.mas.meta ? mainProgram;
+    super.mas.overrideAttrs
+    (old: { meta = old.meta // { mainProgram = "mas"; }; });
+}